refactor: split async runner parallel tests
